refactor: split metric tests
