+++ pageserver/src/consumption_metrics/metrics.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod tests;
